requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
yt-dlp>=2024.0
requests>=2.28
python-dotenv>=1.0
